Cleanup: moved WS in PanierService + badge panier in menu
