+++ src/app/menu/menu.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { SharedService } from '../services/shared/shared.service';
+import { PanierService } from '../services/panier/panier.service';
 
 @Component({
   selector: 'app-menu',
@@ -7,7 +8,10 @@ import { SharedService } from '../services/shared/shared.service';
   styleUrls: ['./menu.component.css']
 })
 export class MenuComponent implements OnInit {
-  constructor(public sharedService: SharedService) { }
+  constructor(
+    public sharedService: SharedService,
+    public panierService: PanierService
+  ) { }
 
   ngOnInit() {
     this.sharedService.isConnected = false;
