+++ src/app/services/shared/shared.service.ts
@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Client } from '../../models/client';
 import { Auteur } from '../../models/auteur';
-import { Panier } from '../../models/panier';
 
 @Injectable()
 export class SharedService {
@@ -23,23 +22,5 @@ export class SharedService {
     }
 
     return url;
-  }
-
-  public setPanier(panier: Panier) {
-    localStorage.setItem('panier', JSON.stringify(panier));
-  }
-
-  public getPanier(): Panier {
-    var panier : Panier = Panier.fromJSON(localStorage.getItem('panier'));
-
-    if (panier == undefined) {
-      return new Panier();
-    } else {
-      return panier;
-    }
-  }
-
-  public clearPanier(): void {
-    localStorage.setItem('panier', JSON.stringify(new Panier()));
   }
 }
